+++ R/algebra.R
@@ -210,20 +210,6 @@ is.algebra <- function(a) {
 }
 
 
-.elementWisely <- function(f) {
-    function(...) {
-        elts <- list(...)
-        if (length(elts) <= 0L) {
-            return(NULL)
-        }
-        vals <- lapply(elts, as.numeric)
-        res <- do.call('mapply', c(list(f), vals))
-        mostattributes(res) <- attributes(elts[[1L]])
-        return(res)
-    }
-}
-
-
 ###########################################################
 # t-norms
 
@@ -250,15 +236,43 @@ goguen.tnorm <- function(...) {
 
 #' @rdname algebra
 #' @export
-pgoedel.tnorm <- .elementWisely(goedel.tnorm)
+pgoedel.tnorm <- function(...) {
+  elts <- list(...)
+  if (length(elts) <= 0L) {
+    return(NULL);
+  }
+  vals <- lapply(elts, as.numeric)
+  res <- .Call('_lfl_pgoedel_tnorm', vals, PACKAGE='lfl')
+  mostattributes(res) <- attributes(elts[[1L]])
+  res
+}
 
 #' @rdname algebra
 #' @export
-plukas.tnorm <- .elementWisely(lukas.tnorm)
+plukas.tnorm <- function(...) {
+  elts <- list(...)
+  if (length(elts) <= 0L) {
+    return(NULL);
+  }
+  vals <- lapply(elts, as.numeric)
+  res <- .Call('_lfl_plukas_tnorm', vals, PACKAGE='lfl')
+  mostattributes(res) <- attributes(elts[[1L]])
+  res
+}
+
 
 #' @rdname algebra
 #' @export
-pgoguen.tnorm <- .elementWisely(goguen.tnorm)
+pgoguen.tnorm <- function(...) {
+  elts <- list(...)
+  if (length(elts) <= 0L) {
+    return(NULL);
+  }
+  vals <- lapply(elts, as.numeric)
+  res <- .Call('_lfl_pgoguen_tnorm', vals, PACKAGE='lfl')
+  mostattributes(res) <- attributes(elts[[1L]])
+  res
+}
 
 
 ###########################################################
@@ -287,15 +301,43 @@ goguen.tconorm <- function(...) {
 
 #' @rdname algebra
 #' @export
-pgoedel.tconorm <- .elementWisely(goedel.tconorm)
+pgoedel.tconorm <- function(...) {
+  elts <- list(...)
+  if (length(elts) <= 0L) {
+    return(NULL);
+  }
+  vals <- lapply(elts, as.numeric)
+  res <- .Call('_lfl_pgoedel_tconorm', vals, PACKAGE='lfl')
+  mostattributes(res) <- attributes(elts[[1L]])
+  res
+}
 
 #' @rdname algebra
 #' @export
-plukas.tconorm <- .elementWisely(lukas.tconorm)
+plukas.tconorm <- function(...) {
+  elts <- list(...)
+  if (length(elts) <= 0L) {
+    return(NULL);
+  }
+  vals <- lapply(elts, as.numeric)
+  res <- .Call('_lfl_plukas_tconorm', vals, PACKAGE='lfl')
+  mostattributes(res) <- attributes(elts[[1L]])
+  res
+}
+
 
 #' @rdname algebra
 #' @export
-pgoguen.tconorm <- .elementWisely(goguen.tconorm)
+pgoguen.tconorm <- function(...) {
+  elts <- list(...)
+  if (length(elts) <= 0L) {
+    return(NULL);
+  }
+  vals <- lapply(elts, as.numeric)
+  res <- .Call('_lfl_pgoguen_tconorm', vals, PACKAGE='lfl')
+  mostattributes(res) <- attributes(elts[[1L]])
+  res
+}
 
 
 ###########################################################