+++ src/RcppExports.cpp
@@ -21,6 +21,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// pgoedel_tnorm
+NumericVector pgoedel_tnorm(List list);
+RcppExport SEXP _lfl_pgoedel_tnorm(SEXP listSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type list(listSEXP);
+    rcpp_result_gen = Rcpp::wrap(pgoedel_tnorm(list));
+    return rcpp_result_gen;
+END_RCPP
+}
 // lukas_tnorm
 double lukas_tnorm(NumericVector vals);
 RcppExport SEXP _lfl_lukas_tnorm(SEXP valsSEXP) {
@@ -29,6 +40,17 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericVector >::type vals(valsSEXP);
     rcpp_result_gen = Rcpp::wrap(lukas_tnorm(vals));
+    return rcpp_result_gen;
+END_RCPP
+}
+// plukas_tnorm
+NumericVector plukas_tnorm(List list);
+RcppExport SEXP _lfl_plukas_tnorm(SEXP listSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type list(listSEXP);
+    rcpp_result_gen = Rcpp::wrap(plukas_tnorm(list));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -43,6 +65,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// pgoguen_tnorm
+NumericVector pgoguen_tnorm(List list);
+RcppExport SEXP _lfl_pgoguen_tnorm(SEXP listSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type list(listSEXP);
+    rcpp_result_gen = Rcpp::wrap(pgoguen_tnorm(list));
+    return rcpp_result_gen;
+END_RCPP
+}
 // goedel_tconorm
 double goedel_tconorm(NumericVector vals);
 RcppExport SEXP _lfl_goedel_tconorm(SEXP valsSEXP) {
@@ -51,6 +84,17 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericVector >::type vals(valsSEXP);
     rcpp_result_gen = Rcpp::wrap(goedel_tconorm(vals));
+    return rcpp_result_gen;
+END_RCPP
+}
+// pgoedel_tconorm
+NumericVector pgoedel_tconorm(List list);
+RcppExport SEXP _lfl_pgoedel_tconorm(SEXP listSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type list(listSEXP);
+    rcpp_result_gen = Rcpp::wrap(pgoedel_tconorm(list));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -65,6 +109,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// plukas_tconorm
+NumericVector plukas_tconorm(List list);
+RcppExport SEXP _lfl_plukas_tconorm(SEXP listSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type list(listSEXP);
+    rcpp_result_gen = Rcpp::wrap(plukas_tconorm(list));
+    return rcpp_result_gen;
+END_RCPP
+}
 // goguen_tconorm
 double goguen_tconorm(NumericVector vals);
 RcppExport SEXP _lfl_goguen_tconorm(SEXP valsSEXP) {
@@ -73,6 +128,17 @@ BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericVector >::type vals(valsSEXP);
     rcpp_result_gen = Rcpp::wrap(goguen_tconorm(vals));
+    return rcpp_result_gen;
+END_RCPP
+}
+// pgoguen_tconorm
+NumericVector pgoguen_tconorm(List list);
+RcppExport SEXP _lfl_pgoguen_tconorm(SEXP listSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type list(listSEXP);
+    rcpp_result_gen = Rcpp::wrap(pgoguen_tconorm(list));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -223,11 +289,17 @@ END_RCPP
 
 static const R_CallMethodDef CallEntries[] = {
     {"_lfl_goedel_tnorm", (DL_FUNC) &_lfl_goedel_tnorm, 1},
+    {"_lfl_pgoedel_tnorm", (DL_FUNC) &_lfl_pgoedel_tnorm, 1},
     {"_lfl_lukas_tnorm", (DL_FUNC) &_lfl_lukas_tnorm, 1},
+    {"_lfl_plukas_tnorm", (DL_FUNC) &_lfl_plukas_tnorm, 1},
     {"_lfl_goguen_tnorm", (DL_FUNC) &_lfl_goguen_tnorm, 1},
+    {"_lfl_pgoguen_tnorm", (DL_FUNC) &_lfl_pgoguen_tnorm, 1},
     {"_lfl_goedel_tconorm", (DL_FUNC) &_lfl_goedel_tconorm, 1},
+    {"_lfl_pgoedel_tconorm", (DL_FUNC) &_lfl_pgoedel_tconorm, 1},
     {"_lfl_lukas_tconorm", (DL_FUNC) &_lfl_lukas_tconorm, 1},
+    {"_lfl_plukas_tconorm", (DL_FUNC) &_lfl_plukas_tconorm, 1},
     {"_lfl_goguen_tconorm", (DL_FUNC) &_lfl_goguen_tconorm, 1},
+    {"_lfl_pgoguen_tconorm", (DL_FUNC) &_lfl_pgoguen_tconorm, 1},
     {"_lfl_goedel_residuum", (DL_FUNC) &_lfl_goedel_residuum, 2},
     {"_lfl_lukas_residuum", (DL_FUNC) &_lfl_lukas_residuum, 2},
     {"_lfl_goguen_residuum", (DL_FUNC) &_lfl_goguen_residuum, 2},