+++ NEWS.md
@@ -1,3 +1,9 @@
+# lfl 2.2.0
+* added equidist(), equifreq()
+* added function to compute fuzzy-transform (ft()) and its inverse (ftinv())
+* parallel t-norms and t-conorms rewritten in C++
+
+
 # lfl 2.1.2
 * added vignette